Reorganizing test structure
