+++ kratos/tests/cpp_tests/core_test_suites.h
@@ -1,0 +1,60 @@
+//    |  /           |
+//    ' /   __| _` | __|  _ \   __|
+//    . \  |   (   | |   (   |\__ `
+//   _|\_\_|  \__,_|\__|\___/ ____/
+//                   Multi-Physics
+//
+//  License:         BSD License
+//                   Kratos default license: kratos/license.txt
+//
+//  Main authors:    Carlos A. Roig
+//
+//
+
+#pragma once 
+
+// System includes
+
+// External includes
+
+// Project includes
+#include "tests/test_utilities/test_suites.h"
+
+// Define some suits that are needed by 
+class KratosSensitivityTestSuite : public KratosCoreFastSuite {};
+class KratosCoreGeometriesFastSuite : public KratosCoreFastSuite {};
+class KratosCoreGeometryContainerFastSuite : public KratosCoreFastSuite {};
+class KratosCoreNurbsGeometriesFastSuite : public KratosCoreFastSuite {};
+class KratosCoreCouplingGeometriesFastSuite : public KratosCoreFastSuite {};
+class KratosExternalLibrariesFastSuite : public KratosCoreFastSuite {};
+class KratosNonRectangularJacobianFastSuite : public KratosCoreFastSuite {};
+class KratosCoreStressSuite : public KratosCoreFastSuite {};
+
+// This classes are temporal and should be changed. Please see: GeoMechanicsApplication, StructuralMechanicsApplication or TrilinosApplication
+// TODO: Remove once every test includes its own suit
+class FluidDynamicsApplicationFastSuite : public KratosCoreFastSuite {};
+class CompressiblePotentialApplicationFastSuite : public KratosCoreFastSuite {};
+class KratosConstitutiveLawsFastSuite : public KratosCoreFastSuite {};
+class KratosContactStructuralMechanicsFastSuite : public KratosCoreFastSuite {};
+class KratosConvectionDiffusionFastSuite : public KratosCoreFastSuite {};
+class KratosCSharpWrapperApplicationFastSuite : public KratosCoreFastSuite {};
+class ExaquteSandboxApplicationFastSuite : public KratosCoreFastSuite {};
+class FluidDynamicsBiomedicalApplicationFastSuite : public KratosCoreFastSuite {};
+class FSIApplicationFastSuite : public KratosCoreFastSuite {};
+class KratosIgaFastSuite : public KratosCoreFastSuite {};
+class KratosIgaFast5PSuite : public KratosCoreFastSuite {};
+class KratosMappingApplicationSerialTestSuite : public KratosCoreFastSuite {};
+class KratosMedFastSuite : public KratosCoreFastSuite {};
+class MeshMovingApplicationFastSuite : public KratosCoreFastSuite {};
+class KratosMeshingApplicationFastSuite : public KratosCoreFastSuite {};
+class KratosRansFastSuite : public KratosCoreFastSuite {};
+class RomApplicationFastSuite : public KratosCoreFastSuite {};
+class KratosSolidMechanicsFastSuite : public KratosCoreFastSuite {};
+class KratosStatisticsFastSuite : public KratosCoreFastSuite {};
+class KratosWindEngineeringFastSuite : public KratosCoreFastSuite {};
+class KratosMPMFastSuite : public KratosCoreFastSuite {};
+class KratosHDF5TestSuite : public KratosCoreFastSuite {};
+class ShallowWaterApplicationFastSuite : public KratosCoreFastSuite {};
+
+// TODO: those should be in its own mpi suite
+class KratosMappingApplicationMPITestSuite : public KratosCoreFastSuite {};
