+++ kratos/tests/test_utilities/runtime_dependency_handler.cpp
@@ -1,3 +1,16 @@
+//    |  /           |
+//    ' /   __| _` | __|  _ \   __|
+//    . \  |   (   | |   (   |\__ `
+//   _|\_\_|  \__,_|\__|\___/ ____/
+//                   Multi-Physics
+//
+//  License:         BSD License
+//                   Kratos default license: kratos/license.txt
+//
+//  Main authors:    Jordi Cotela Dalmau
+//
+//
+
 #include "runtime_dependency_handler.h"
 
 namespace Kratos::Testing {
